finrise/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==3.0.4
channels-redis==4.2.0
constantly==23.10.4
cryptography==41.0.7
daphne==3.0.2
Django==4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gevent==24.2.1
greenlet==3.0.3
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytz==2024.1
pywin32==225
redis==5.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing-extensions==4.10.0
tzdata==2024.1
websocket==0.2.1
websockets==12.0
zope.event==5.0
zope.interface==6.2